+++ TonyaPuzzle/script/component/tp-topnav/etopnav.js
@@ -1,4 +1,4 @@
 export * from "./tp-topnav.js";
 
 export * from "./tp-topnav-item.js";
-export * from "./tp-avatar.js";
+export * from "./tp-topnav-avater.js";