+++ TonyaPuzzle/tp.js
@@ -1,10 +1,8 @@
-// export * from "./script/component/tp-topnav/tp-topnav.js";
-
 // topnav
 export * from "./script/component/tp-topnav/etopnav.js"
 // sidebar
-export * from "./script/component/tp-sidebar.js";
+export * from "./script/component/tp-sidebar/esidebar.js";
 // main
-export * from "./script/component/tp-main.js";
-
-export * from "./script/component/tp-card.js";
+export * from "./script/component/tp-main/emain.js";
+// common
+export * from "./script/component/tp-common/ecommon.js";